Rename abstract_processor_step.py to processor_step.py
